Add DOUBLE_FILLET_GEOM_BASED, move functions out of bc_colloc.c
