+++ include/bc_colloc.h
@@ -59,56 +59,6 @@ apply_point_colloc_bc(double[],                     /* resid_vector */
                       const double,                 /* time value */
                       Exo_DB *);
 
-EXTERN void moving_plane(int,      /* ielem_dim */
-                         double *, /* func */
-                         double[], /* d_func */
-                         dbl *,    /* aa */
-                         double);  /* time */
-
-EXTERN void fmesh_constraint(double *,   /* func */
-                             double[],   /* d_func */
-                             const int); /* bc_id */
-
-EXTERN void fplane(const int, /* ielem_dim */
-                   double *,  /* func */
-                   double[],  /* d_func - dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                   dbl *);    /* aa - function parameters from data card  */
-
-EXTERN void f_fillet(const int,      /* ielem_dim */
-                     double *,       /* func */
-                     double[],       /* d_func - dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                     const double *, /* p - function parameters from data card  */
-                     const int);     /* number of parameters from bc card  */
-
-EXTERN void f_double_rad(const int,      /* ielem_dim */
-                         double *,       /* func */
-                         double[],       /* d_func - dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                         const double *, /* p - function parameters from data card  */
-                         const int);     /* number of parameters from bc card  */
-
-void f_double_fillet(const int ielem_dim,
-                     double *func,
-                     double d_func[],      /* dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                     const double *p,      /*  function parameters from data card  */
-                     const int num_const); /* number of passed parameters   */
-
-#ifdef FEATURE_ROLLON_PLEASE
-EXTERN void f_feature_rollon(const int, /* ielem_dim */
-                             double *,  /* func */
-                             double[],  /* d_func - dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                             const double *, /* p - function parameters from data card  */
-                             const int,      /* number of parameters from bc card  */
-                             const int,      /* geometry model id  */
-                             const double);  /* time - time at which BC's are evaluated  */
-#endif
-
-EXTERN void f_roll_fluid(const int,      /* ielem_dim */
-                         double *,       /* func */
-                         double[],       /* d_func - dimensioned [MAX_VARIABLE_TYPES+MAX_CONC] */
-                         const double *, /* p - function parameters from data card  */
-                         const int,      /* number of parameters from bc card  */
-                         double *);      /* number of parameters from bc card  */
-
 EXTERN void fvelocity_profile(const int,     /* var_flag */
                               const int,     /* ielem_dim */
                               const int,     /* velo_condition */
@@ -136,74 +86,8 @@ EXTERN void f_vestress_parabola(const int,      /* var_flag */
                                 const double,   /* time - time at which BC's are evaluated  */
                                 const int);     /* number of parameters */
 
-EXTERN void fspline(const int,     /* ielem_dim */
-                    double *,      /* func */
-                    double[],      /* d_func - [MAX_VARIABLE_TYPES + MAX_CONC] */
-                    double[],      /* p - parameterize eqn model */
-                    const double); /* time - at which bc's are evaluated */
-
-EXTERN void fspline_rs(const int,     /* ielem_dim */
-                       double *,      /* func */
-                       double[],      /* d_func - [MAX_VARIABLE_TYPES + MAX_CONC] */
-                       double[],      /* p - parameterize eqn model */
-                       const double); /* time - at which bc's are evaluated */
-
 EXTERN void fTmelting(double *, /* func */
                       double[], /* d_func - [MAX_VARIABLE_TYPES + MAX_CONC] */
                       double);  /* a1 - function parameter from data card   */
-
-EXTERN int fgeneralized_dirichlet(double *,      /* func */
-                                  double[],      /* d_func - MAX_VARIABLE_TYPES + MAX_CONC */
-                                  const int,     /* gd_condition - denoting which condition
-                                                  * applied */
-                                  const int,     /* bc_input_id */
-                                  const double,  /* tt - parameter to vary time integration
-                                                  * from explicit (tt = 1) to
-                                                  * implicit (tt = 0) */
-                                  const double); /* dt - current time step size          */
-
-EXTERN int load_variable(double *,     /* x_var - variable value */
-                         double *,     /* d_x_var - sensitivities of variable value */
-                         const int,    /* jvar - variable number */
-                         const int,    /* wspec - species number */
-                         const double, /* tt - parameter to vary time integration
-                                        * from explicit (tt = 1) to
-                                        * implicit (tt = 0) */
-                         const double, /* dt - current time step size */
-                         double[]);    /* vector sensitivity vector - SPEED */
-
-extern int bc_eqn_index(int, int, int, int, int, int *, int *, VARIABLE_DESCRIPTION_STRUCT **);
-
-EXTERN int bc_eqn_index_stress(int id,          /* local node number                 */
-                               int I,           /* processor node number             */
-                               int bc_input_id, /* boundary condition number         */
-                               int curr_matID,  /* Current material ID */
-                               int kdir,        /* coordinate index for stress components */
-                               int mode,        /* Stress mode number */
-                               int *eqn,        /* eqn to which this condition is applied     */
-                               int *matID_retn, /* material ID to apply this eqn on           */
-                               VARIABLE_DESCRIPTION_STRUCT **vd_retn);
-
-EXTERN int evaluate_time_func(const double, /* time                                      */
-                              double *,     /* f_time - computed time function           */
-                              const int);   /* bc_input_id                               */
-
-EXTERN void apply_table_bc(double *, /* func                                      */
-                           double[MAX_VARIABLE_TYPES + MAX_CONC], /* d_func                       */
-                           struct Boundary_Condition *, /* BC_Type                             */
-                           double);                     /* time _value */
-
-EXTERN double interpolate_table(struct Data_Table *, /* table               */
-                                double[],            /* x            */
-                                double *,            /* slope                 */
-                                double[]);           /* gradient array         */
-
-EXTERN double table_distance_search(struct Data_Table *, /* table               */
-                                    double[],            /* x            */
-                                    double *,            /* slope                 */
-                                    double[]);           /* gradient array         */
-
-EXTERN double interpolate_table_sat(struct Data_Table *, /* table */
-                                    double[DIM]); /* slope                                     */
 
 #endif /* GOMA_BC_COLLOC_H */
