+++ include/mm_names.h
@@ -467,6 +467,16 @@ struct BC_descriptions BC_Desc[] = {
      {0, 0, 0, 0, 0, 1, 1, 1, 0, 0},
      SINGLE_PHASE,
      DVI_SINGLE_PHASE_DB},
+    {"DOUBLE_FILLET_GEOM_BASED",
+     "DOUBLE_FILLET_GEOM_BASED_BC",
+     COLLOCATE_SURF,
+     DOUBLE_FILLET_GEOM_BASED_BC,
+     R_MESH_NORMAL,
+     SCALAR,
+     R_MESH1,
+     {0, 0, 0, 0, 0, 1, 1, 1, 0, 0},
+     SINGLE_PHASE,
+     DVI_SINGLE_PHASE_DB},
     {"FEATURE_ROLLON",
      "FEATURE_ROLLON_BC",
      COLLOCATE_SURF,
