+++ include/rf_bc_const.h
@@ -791,108 +791,109 @@
 #define SPLINEZ_RS_BC 932000000
 #define SPLINE_RS_BC  931100000
 /* replace SPLINE boundary conditions by GEOM */
-#define GEOMX_BC                   910000000
-#define GEOMY_BC                   920000000
-#define GEOMZ_BC                   930000000
-#define GEOM_BC                    931000000
-#define PLANEX_BC                  940000000
-#define PLANEY_BC                  950000000
-#define PLANEZ_BC                  960000000
-#define PLANE_BC                   961000000
-#define FILLET_BC                  961123400
-#define DOUBLE_RAD_BC              961123500
-#define DOUBLE_FILLET_BC           961123501
-#define FEATURE_ROLLON_BC          961223400
-#define ROLL_FLUID_BC              961124500
-#define SA_WALL_FUNC_BC            961124501
-#define OMEGA_WALL_FUNC_BC         961124502
-#define FLUIDITY_EQUILIBRIUM_BC    961124503
-#define TENSION_SHEET_BC           96210200
-#define MOVING_PLANE_BC            96110000
-#define MOVING_PLANE_ETCH_BC       96115000
-#define SM_PLANE_BC                961200000 /* Solid Model PLANE BC */
-#define MESH_CONSTRAINT_BC         961300000
-#define KINEMATIC_BC               962000000
-#define KINEMATIC_XI_BC            962000001
-#define KINEMATIC_ETA_BC           962000002
-#define LUB_KINEMATIC_BC           962050000
-#define KIN_LEAK_BC                962100000
+#define GEOMX_BC                    910000000
+#define GEOMY_BC                    920000000
+#define GEOMZ_BC                    930000000
+#define GEOM_BC                     931000000
+#define PLANEX_BC                   940000000
+#define PLANEY_BC                   950000000
+#define PLANEZ_BC                   960000000
+#define PLANE_BC                    961000000
+#define FILLET_BC                   961123400
+#define DOUBLE_RAD_BC               961123500
+#define DOUBLE_FILLET_BC            961123501
+#define DOUBLE_FILLET_GEOM_BASED_BC 961123502
+#define FEATURE_ROLLON_BC           961223400
+#define ROLL_FLUID_BC               961124500
+#define SA_WALL_FUNC_BC             961124501
+#define OMEGA_WALL_FUNC_BC          961124502
+#define FLUIDITY_EQUILIBRIUM_BC     961124503
+#define TENSION_SHEET_BC            96210200
+#define MOVING_PLANE_BC             96110000
+#define MOVING_PLANE_ETCH_BC        96115000
+#define SM_PLANE_BC                 961200000 /* Solid Model PLANE BC */
+#define MESH_CONSTRAINT_BC          961300000
+#define KINEMATIC_BC                962000000
+#define KINEMATIC_XI_BC             962000001
+#define KINEMATIC_ETA_BC            962000002
+#define LUB_KINEMATIC_BC            962050000
+#define KIN_LEAK_BC                 962100000
 #define KIN_ANTOINE_BC             962150000
-#define KIN_ELECTRODEPOSITION_BC   962200000 /* RSL 5/24/02 */
-#define KINEMATIC_EDGE_BC          962400000
-#define KIN_CHEM_BC                962500000
-#define KIN_LEAK_HEAT_BC           962600000
-#define CAPILLARY_BC               963000000
-#define CAPILLARY_SHEAR_VISC_BC    963000002
-#define CAPILLARY_TABLE_BC         963000007
-#define ELEC_TRACTION_BC           963000001 /* Include Maxwell Stress in CAPILLARY_BC */
-#define ELEC_TRACTION_SOLID_BC     963000003 /* Include Maxwell Stress in solid */
-#define SHEAR_STRESS_APPLIED_BC    963000009
-#define CAP_REPULSE_BC             963100000
-#define CAP_RECOIL_PRESS_BC        963110000
-#define CAP_REPULSE_ROLL_BC        963120000
-#define CAP_REPULSE_USER_BC        963130000
-#define CAP_REPULSE_TABLE_BC       963140000
-#define PRESSURE_USER_BC           963200000
-#define FLOW_PRESSURE_BC           963300000
-#define FLOW_HYDROSTATIC_BC        963400000
-#define FLOW_PRESSURE_VAR_BC       963400001
-#define FLOW_PRESS_USER_BC         963500000
-#define HYDROSTATIC_SYMM_BC        963600000
-#define FLOW_STRESSNOBC_BC         963700000
-#define FLOW_GRADV_BC              963800000
-#define FLOW_GRADV_SIC_BC          963800001
-#define FLOW_GRADV_T_BC            963800002
-#define SHEET_ENDSLOPE_BC          963900000
-#define CA_BC                      964000000
-#define CA_MOMENTUM_BC             964000008
-#define CA_EDGE_BC                 964000001
-#define CA_EDGE_INT_BC             964000002
-#define CA_OR_FIX_BC               964000003
-#define VAR_CA_EDGE_BC             964000004
-#define VAR_CA_USER_BC             964000005
-#define CA_EDGE_CURVE_BC           964000006
-#define CA_EDGE_CURVE_INT_BC       964000007
-#define SURFTANG_BC                964100000
-#define SURFTANG_SCALAR_BC         964110000
-#define CAP_ENDFORCE_BC            964100010
-#define CAP_ENDFORCE_SCALAR_BC     964110010
-#define SURFTANG_EDGE_BC           964100001
-#define SURFTANG_SCALAR_EDGE_BC    964100002
-#define VELO_NORMAL_BC             964200000
-#define VELO_NORMAL_EDGE_INT_BC    964200001
-#define VELO_NORMAL_EDGE_BC        964200002
-#define VELO_NORM_COLLOC_BC        964200003
-#define VELO_TANG1_COLLOC_BC       964200010
-#define VELO_TANG2_COLLOC_BC       964200011
-#define VELO_NORMAL_SOLID_BC       964200004
-#define VELO_NORMAL_LS_BC          964200005
-#define VELO_NORMAL_LS_COLLOC_BC   964200006
-#define VELO_NORMAL_LS_PETROV_BC   964200007
-#define VELO_NORMAL_LUB_BC         964200008
-#define VNORM_LEAK_BC              964210000
-#define VNORM_ELECTRODEPOSITION_BC 964220000 /* RSL 5/30/02 */
-#define VELO_TANGENT_BC            964300000
-#define VELO_TANGENT_EDGE_INT_BC   964300001
-#define VELO_TANGENT_EDGE_BC       964300002
-#define VELO_TANGENT_SOLID_BC      964300003
-#define VELO_TANGENT_3D_BC         964300004
-#define VELO_TANGENT_USER_BC       964300005
-#define VELO_TANGENT_LS_BC         964300006
-#define ZERO_VELO_TANGENT_3D_BC    964300007
-#define VELO_SLIP_BC               964400000
-#define VELO_SLIP_ROT_BC           964410000
-#define VELO_SLIP_FILL_BC          964420000
-#define VELO_SLIP_LEVEL_BC         964421000
-#define VELO_SLIP_LEVEL_SIC_BC     964421010
-#define VELO_SLIP_LS_ROT_BC        964421328
-#define VELO_SLIP_LS_HEAVISIDE_BC  964421009
-#define VELO_SLIP_LS_ORIENTED_BC   964421011
-#define VELO_SLIP_SOLID_BC         964430000
-#define VELO_SLIP_ROT_FILL_BC      964435000
-#define VELO_SLIP_EK_BC            964440000
-#define VELO_SLIP_POWER_BC         964400001
-#define VELO_SLIP_POWER_CARD_BC    964400002
+#define KIN_ELECTRODEPOSITION_BC    962200000 /* RSL 5/24/02 */
+#define KINEMATIC_EDGE_BC           962400000
+#define KIN_CHEM_BC                 962500000
+#define KIN_LEAK_HEAT_BC            962600000
+#define CAPILLARY_BC                963000000
+#define CAPILLARY_SHEAR_VISC_BC     963000002
+#define CAPILLARY_TABLE_BC          963000007
+#define ELEC_TRACTION_BC            963000001 /* Include Maxwell Stress in CAPILLARY_BC */
+#define ELEC_TRACTION_SOLID_BC      963000003 /* Include Maxwell Stress in solid */
+#define SHEAR_STRESS_APPLIED_BC     963000009
+#define CAP_REPULSE_BC              963100000
+#define CAP_RECOIL_PRESS_BC         963110000
+#define CAP_REPULSE_ROLL_BC         963120000
+#define CAP_REPULSE_USER_BC         963130000
+#define CAP_REPULSE_TABLE_BC        963140000
+#define PRESSURE_USER_BC            963200000
+#define FLOW_PRESSURE_BC            963300000
+#define FLOW_HYDROSTATIC_BC         963400000
+#define FLOW_PRESSURE_VAR_BC        963400001
+#define FLOW_PRESS_USER_BC          963500000
+#define HYDROSTATIC_SYMM_BC         963600000
+#define FLOW_STRESSNOBC_BC          963700000
+#define FLOW_GRADV_BC               963800000
+#define FLOW_GRADV_SIC_BC           963800001
+#define FLOW_GRADV_T_BC             963800002
+#define SHEET_ENDSLOPE_BC           963900000
+#define CA_BC                       964000000
+#define CA_MOMENTUM_BC              964000008
+#define CA_EDGE_BC                  964000001
+#define CA_EDGE_INT_BC              964000002
+#define CA_OR_FIX_BC                964000003
+#define VAR_CA_EDGE_BC              964000004
+#define VAR_CA_USER_BC              964000005
+#define CA_EDGE_CURVE_BC            964000006
+#define CA_EDGE_CURVE_INT_BC        964000007
+#define SURFTANG_BC                 964100000
+#define SURFTANG_SCALAR_BC          964110000
+#define CAP_ENDFORCE_BC             964100010
+#define CAP_ENDFORCE_SCALAR_BC      964110010
+#define SURFTANG_EDGE_BC            964100001
+#define SURFTANG_SCALAR_EDGE_BC     964100002
+#define VELO_NORMAL_BC              964200000
+#define VELO_NORMAL_EDGE_INT_BC     964200001
+#define VELO_NORMAL_EDGE_BC         964200002
+#define VELO_NORM_COLLOC_BC         964200003
+#define VELO_TANG1_COLLOC_BC        964200010
+#define VELO_TANG2_COLLOC_BC        964200011
+#define VELO_NORMAL_SOLID_BC        964200004
+#define VELO_NORMAL_LS_BC           964200005
+#define VELO_NORMAL_LS_COLLOC_BC    964200006
+#define VELO_NORMAL_LS_PETROV_BC    964200007
+#define VELO_NORMAL_LUB_BC          964200008
+#define VNORM_LEAK_BC               964210000
+#define VNORM_ELECTRODEPOSITION_BC  964220000 /* RSL 5/30/02 */
+#define VELO_TANGENT_BC             964300000
+#define VELO_TANGENT_EDGE_INT_BC    964300001
+#define VELO_TANGENT_EDGE_BC        964300002
+#define VELO_TANGENT_SOLID_BC       964300003
+#define VELO_TANGENT_3D_BC          964300004
+#define VELO_TANGENT_USER_BC        964300005
+#define VELO_TANGENT_LS_BC          964300006
+#define ZERO_VELO_TANGENT_3D_BC     964300007
+#define VELO_SLIP_BC                964400000
+#define VELO_SLIP_ROT_BC            964410000
+#define VELO_SLIP_FILL_BC           964420000
+#define VELO_SLIP_LEVEL_BC          964421000
+#define VELO_SLIP_LEVEL_SIC_BC      964421010
+#define VELO_SLIP_LS_ROT_BC         964421328
+#define VELO_SLIP_LS_HEAVISIDE_BC   964421009
+#define VELO_SLIP_LS_ORIENTED_BC    964421011
+#define VELO_SLIP_SOLID_BC          964430000
+#define VELO_SLIP_ROT_FILL_BC       964435000
+#define VELO_SLIP_EK_BC             964440000
+#define VELO_SLIP_POWER_BC          964400001
+#define VELO_SLIP_POWER_CARD_BC     964400002
 
 #define VELO_EK_3D_BC          964450000
 #define MOVING_CA_BC           964500000
