+++ src/bc_integ.c
@@ -31,6 +31,7 @@
 
 #include "ac_stability.h"
 #include "ac_stability_util.h"
+#include "bc/eqn_index.h"
 #include "bc/rotate_coordinates.h"
 #include "bc_colloc.h"
 #include "bc_integ.h"
@@ -77,6 +78,7 @@
 #include "rf_solver.h"
 #include "rf_vars_const.h"
 #include "std.h"
+#include "table.h"
 #include "user_bc.h"
 #include "util/goma_normal.h"
