+++ src/mm_augc_util.c
@@ -156,6 +156,7 @@ void load_extra_unknownsAC(int iAC,     /* ID NUMBER OF AC'S */
       case FILLET_BC:
       case DOUBLE_RAD_BC:
       case DOUBLE_FILLET_BC:
+      case DOUBLE_FILLET_GEOM_BASED_BC:
       case FEATURE_ROLLON_BC:
       case ROLL_FLUID_BC:
       case UVARY_BC:
@@ -922,6 +923,7 @@ void update_parameterAC(
           case FILLET_BC:
           case DOUBLE_RAD_BC:
           case DOUBLE_FILLET_BC:
+          case DOUBLE_FILLET_GEOM_BASED_BC:
           case FEATURE_ROLLON_BC:
           case ROLL_FLUID_BC:
           case UVARY_BC:
@@ -991,6 +993,7 @@ void update_parameterAC(
       case FILLET_BC:
       case DOUBLE_RAD_BC:
       case DOUBLE_FILLET_BC:
+      case DOUBLE_FILLET_GEOM_BASED_BC:
       case FEATURE_ROLLON_BC:
       case ROLL_FLUID_BC:
       case UVARY_BC:
