+++ src/mm_fill_util.c
@@ -26,6 +26,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "bc/eqn_index.h"
 #include "bc/rotate_coordinates.h"
 #include "bc_contact.h"
 #include "dp_utils.h"
