+++ src/mm_flux.c
@@ -18,6 +18,7 @@
 
 /* Standard include files */
 #include "load_field_variables.h"
+#include "table.h"
 #include <complex.h>
 #undef I
 #include <math.h>
